refactor: rename collections, datasets, genesets files
