+++ backend/portal/api/app/portal-api.yml
@@ -24,7 +24,7 @@ paths:
       description: >-
         This lists all collections and their IDs that currently exist in the corpora. If a parameter is specified as
         a filter, then only collections that meet the status criteria will be outputted.
-      operationId: backend.portal.api.app.v1.collection.get_collections_list
+      operationId: backend.portal.api.app.v1.collections.get_collections_list
       parameters:
         - $ref: "#/components/parameters/query_user_id"
         - name: from_date
@@ -73,7 +73,7 @@ paths:
         - cxguserCookie: []
       description: >-
         Creates a new collection in corpora. Once a collection is created it can be linked to datasets.
-      operationId: backend.portal.api.app.v1.collection.create_collection
+      operationId: backend.portal.api.app.v1.collections.create_collection
       requestBody:
         content:
           application/json:
@@ -128,7 +128,7 @@ paths:
         - {}
       description: >-
         This will return all datasets and associated attributes.
-      operationId: backend.portal.api.app.v1.collection.get_collection_details
+      operationId: backend.portal.api.app.v1.collections.get_collection_details
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
       responses:
@@ -156,7 +156,7 @@ paths:
       description: >-
         This will return a private version of the public collection with all datasets and assoicated attributes of the
         public collection.
-      operationId: backend.portal.api.app.v1.collection.post_collection_revision
+      operationId: backend.portal.api.app.v1.collections.post_collection_revision
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
       responses:
@@ -183,7 +183,7 @@ paths:
         - cxguserCookie: []
       description: >-
         Update a private collection's summary, links, contact info, and description details.
-      operationId: backend.portal.api.app.v1.collection.update_collection
+      operationId: backend.portal.api.app.v1.collections.update_collection
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
       requestBody:
@@ -230,7 +230,7 @@ paths:
       description: >-
         Deletes an entire private collection from cellxgene data-portal, including any generated artifacts/assets and
         deployments.
-      operationId: backend.portal.api.app.v1.collection.delete_collection
+      operationId: backend.portal.api.app.v1.collections.delete_collection
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
       responses:
@@ -424,7 +424,7 @@ paths:
       summary: >-
         Returns all public, non tombstoned collections. This endpoint is meant to be used for displaying collections
         in a compact view. It doesn't return all the fields, so it is not suitable for general usage.
-      operationId: backend.portal.api.app.v1.collection.get_collections_index
+      operationId: backend.portal.api.app.v1.collections.get_collections_index
       responses:
         "200":
           description: OK
@@ -460,7 +460,7 @@ paths:
         Cancels the in process upload of a dataset to the data portal and cleans up (removes) any artifacts created
         in the upload process. If the dataset has already been uploaded and converted, this endpoint deletes it from
         the data portal and removes associated artifacts
-      operationId: backend.portal.api.app.v1.dataset.delete_dataset
+      operationId: backend.portal.api.app.v1.datasets.delete_dataset
       security:
         - cxguserCookie: []
       parameters:
@@ -484,7 +484,7 @@ paths:
         Given a cellxgene Explorer url return some basic metadata about the dataset it represent in the cellxgene
         Portal data model including its associated dataset id, the s3_uri of the dataset, its tombstone status and
         the associated parent collection ID.
-      operationId: backend.portal.api.app.v1.dataset.get_dataset_identifiers
+      operationId: backend.portal.api.app.v1.datasets.get_dataset_identifiers
       parameters:
         - in: query
           required: true
@@ -511,7 +511,7 @@ paths:
       summary: Dataset assets
       description: >-
         Return a list of all the dataset's assets
-      operationId: backend.portal.api.app.v1.dataset.get_dataset_assets
+      operationId: backend.portal.api.app.v1.datasets.get_dataset_assets
       parameters:
         - $ref: "#/components/parameters/path_dataset_id"
       responses:
@@ -535,7 +535,7 @@ paths:
       summary: Request to download a dataset
       description: >-
         Request to download a file which on success will generate a pre-signed URL to download the dataset.
-      operationId: backend.portal.api.app.v1.dataset.post_dataset_asset
+      operationId: backend.portal.api.app.v1.datasets.post_dataset_asset
       parameters:
         - $ref: "#/components/parameters/path_dataset_id"
         - name: asset_id
@@ -577,7 +577,7 @@ paths:
       description: >-
         Request to update (add or remove) linkages between a dataset and gene sets. The request must be made by the
         collection owner and the dataset and gene sets must all belong to the same private collection
-      operationId: backend.portal.api.app.v1.dataset.post_dataset_gene_sets
+      operationId: backend.portal.api.app.v1.datasets.post_dataset_gene_sets
       parameters:
         - $ref: "#/components/parameters/path_dataset_id"
       requestBody:
@@ -616,7 +616,7 @@ paths:
       tags:
         - datasets
       summary: Get the status of a dataset upload.
-      operationId: backend.portal.api.app.v1.dataset.get_status
+      operationId: backend.portal.api.app.v1.datasets.get_status
       security:
         - cxguserCookie: []
       parameters:
@@ -711,7 +711,7 @@ paths:
       description: >-
         Returns all datasets that belong to public, non tombstoned collections. This endpoint is meant to be used 
         for displaying datasets in a compact view. It doesn't return all the fields, so it is not suitable for general usage.
-      operationId: backend.portal.api.app.v1.dataset.get_datasets_index
+      operationId: backend.portal.api.app.v1.datasets.get_datasets_index
       responses:
         "200":
           description: OK
@@ -790,7 +790,7 @@ paths:
         - cxguserCookie: []
       description: >-
         Delete a gene by set removing it from a collection and any data sets it was linked to.
-      operationId: backend.portal.api.app.v1.gene_sets.delete
+      operationId: backend.portal.api.app.v1.genesets.delete
       parameters:
         - $ref: "#/components/parameters/path_geneset_id"
       responses:
