+++ tests/unit/backend/api_server/test_v1_collection.py
@@ -16,7 +16,7 @@ from backend.common.corpora_orm import (
 from backend.common.entities import Collection
 from backend.common.providers.crossref_provider import CrossrefDOINotFoundException, CrossrefFetchException
 from backend.common.utils.corpora_constants import CorporaConstants
-from backend.portal.api.app.v1.collection import verify_collection_body
+from backend.portal.api.app.v1.collections import verify_collection_body
 from tests.unit.backend.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
 from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
